rename : 이메일 템플릿 계층 분리 및 패키지 구조 정리
