+++ src/main/java/com/icestudyroom_email/domain/email/infrastructure/template/Top5EnterEmailTemplate.java
@@ -1,4 +1,4 @@
-package com.icestudyroom_email.domain.email.infrastructure.kafka.dto.ranking;
+package com.icestudyroom_email.domain.email.infrastructure.template;
 
 import com.icestudyroom_email.domain.email.infrastructure.gmail.dto.EmailRequest;
 import com.icestudyroom_email.domain.rankingContract.RankingEmailEvent;
